// pop-web/src/app/core/modules/openapi/model/risk-assessment-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CodedConceptSchema } from './coded-concept-schema';


export interface RiskAssessmentCreate { 
    /**
     * Indicates the case of the patient who\'s cancer risk is assesed
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the risk assessment was performed and recorded.
     */
    date: string;
    /**
     * Indicates the method or type of risk assessment
     */
    methodology: CodedConceptSchema;
    /**
     * Assessed risk
     */
    risk: CodedConceptSchema;
    score?: number | null;
    /**
     * References to the neoplastic entities that were assessed to estimate the risk.
     */
    assessedEntitiesIds?: Array<string>;
}